add end_handler hack
This "fixes" a problem when a http[s] request with content doesn't
trigger an 'end' but only a 'close' event.

This problem was detected first when switching from node 0.4.2 to 0.4.8
and seems to exist in 0.4.9, too. At some point we've to fix node but
now it just has to work.^_^